requirements.txt
black == 22.3.0
click == 8.0.4
fastapi >= 0.75.1
fastapi_health >= 0.4.0
isort >= 5.10.1
matplotlib >= 3.5.1
numpy >= 1.21.5
pandas >= 1.3.5
pre_commit >= 2.18.1
PyYAML >= 5.3.1
scikit-learn >= 1.0.2
uvicorn >= 0.17.6
